Code/backend/requirements.txt
setuptools
flask
Flask-SQLAlchemy
flask-restful
Flask-Security-Too
bcrypt==4.0.1
flask-cors